Split up integration tests
Separate them into those that do not need to run a JVM, using
the newly added `java` script, and those that really need to run
the sbt-launch.jar
